pyproject.toml
[build-system]
requires = ['poetry-core']
build-backend = 'poetry.core.masonry.api'

[tool.black]
extend-exclude = '''.protos/.*'''
include = '\.pyi?$'
line-length = 100
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
exclude = [
  'src/protos'
]
ignore_missing_imports = true

[tool.poetry]
authors = [
  'Sitt Guruvanich <aekazitt+github@gmail.com>'
]
description = 'Lightning Service Provider built on top of FastAPI asynchronous web framework'
license = 'MIT'
name = 'tesla-ball'
packages = [
  {include='protoc.py'},
  {include='migrate.py'},
  {include='serve.py'},
  {include='src'}
]
readme = 'README.md'
version = '0.1.0'

[tool.poetry.dependencies]
apscheduler = '^3.10.4'
fastapi = '^0.103.2'
fastapi-csrf-protect = '^0.3.2'
googleapis-common-protos = '1.60.0'
grpcio = '^1.48.0'
grpcio-tools = '^1.48.0'
orjson = '^3.9.7'
pendulum = '^2.1.2'
python = '^3.9'
python-bitcoinlib = '^0.12.2'
pyyaml = '^6.0.1'
schedule = '^1.2.1'
tortoise-orm = {extras = ['asyncpg'], version = '^0.20.0'}
uvicorn = '^0.23.2'

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
aerich = '^0.7.2'
httpx = '^0.25.0'
mypy = '^1.5.1'
pytest = '^7.4.2'
pytest-asyncio = '^0.21.1'
python-dotenv = '^1.0.0'
ruff = '^0.1.14'

[tool.poetry.scripts]
compile-protos = 'protoc:main'
migrate = 'migrate:main'

[tool.pytest.ini_options]
addopts = '--strict-markers --tb=short -s'
python_files = '*.py'
testpaths = [ 'tests' ]


[tool.ruff]
exclude = ['src/protos/*.py']
indent-width = 2
line-length = 100
target-version = 'py39'


[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"] # Ignore unused imports
